MNT: Rename benthic_data_classes.definitions -> benthic_data_classes.datasets
